fix по Критериям
