+++ project/src/store/api-action.ts
@@ -11,9 +11,8 @@ import {
   setOffersLoadingStatus,
   setCommentLoadingStatus
 } from './action';
-import {AuthData, Comment, Offer, Review, UserData} from '../types/types';
+import {AuthData, Comment, Offer, Review, State, UserData} from '../types/types';
 import {APIRoute, AppRoute} from '../const/const';
-import {State} from './reducer';
 import {saveToken} from '../services/token';
 import {generatePath} from 'react-router-dom';
 import {toast} from 'react-toastify';
